lunabot_nonebot/pyproject.toml
[project]
name = "lunabot-nonebot"
version = "0.1.0"
description = "luanbot的docker版"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp",
    "aiorpcx>=0.25.0",
    "aiosmtplib",
    "aiosqlite",
    "asteval>=1.0.8",
    "bs4",
    "colour-science",
    "emoji==2.0.0",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.56.0",
    "imageio[ffmpeg]>=2.37.2",
    "jieba>=0.42.1",
    "matplotlib",
    "memray>=1.19.1",
    "mistune",
    "nonebot-adapter-onebot",
    "nonebot-plugin-apscheduler",
    "nonebot-plugin-htmlrender",
    "nonebot-plugin-picstatus",
    "nonebot2[fastapi]",
    "openai>=2.14.0",
    "opencv-python-headless",
    "orjson",
    "pandas",
    "pdf2image>=1.17.0",
    "picimagesearch>=3.12.11",
    "pillow",
    "pilmoji[requests]==2.0.4",
    "pydub",
    "python-dotenv",
    "pyyaml",
    "pyzbar>=0.1.9",
    "qrcode>=8.2",
    "rapidfuzz",
    "rembg>=2.0.69",
    "rembg>=2.0.69",
    "requests",
    "scikit-learn>=1.8.0",
    "scipy",
    "sekai-deck-recommend-cpp",
    "setproctitle",
    "svgwrite>=1.4.3",
    "tenacity",
    "tencentcloud-sdk-python>=3.1.15",
    "uvloop",
    "wordcloud",
    "yappi",
    "yt-dlp",
    "zhconv>=1.4.3",
    "zhon",
    "zstandard",
]
# 配置镜像源
[[tool.uv.index]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
default = true

# nonebot相关配置
[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_picstatus"]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.pytest.ini_options]
# Add the path to the directory containing the external test files.
# This tells pytest not to look inside these folders for tests.
norecursedirs = [
    ".git",
    ".venv",
    "sekai-deck-recommend-cpp",  # Add this line to exclude the problematic directory
    # or be more specific:
    # "sekai-deck-recommend-cpp/3rdparty"
]
